requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
google-search-results==2.4.2
fpdf==1.7.2
click==8.1.7